Split useInnerWatch/watching-single-store.spec
