+++ tests/useInnerWatch.spec.ts
@@ -25,663 +25,6 @@ interface WithSpy {
   spy: VoidFunction
 }
 
-interface WithInitial<T = Counter> {
-  initial?: T
-}
-
-describe("single store is watching", () => {
-  describe.each([
-    [
-      "inline watcher",
-      ({ store }: WithStore) => {
-        return useInnerWatch(() => store.getState())
-      },
-    ],
-    [
-      "inline watcher with inline comparator",
-      ({ store }: WithStore) => {
-        return useInnerWatch(
-          () => store.getState(),
-          (prev, next) => Counter.compare(prev, next),
-        )
-      },
-    ],
-    [
-      "inline watcher with memoized comparator",
-      ({ store }: WithStore) => {
-        return useInnerWatch(() => store.getState(), Counter.compare)
-      },
-    ],
-    [
-      "memoized watcher",
-      ({ store }: WithStore) => {
-        return useInnerWatch(useCallback(() => store.getState(), [store]))
-      },
-    ],
-    [
-      "memoized watcher with inline comparator",
-      ({ store }: WithStore) => {
-        return useInnerWatch(
-          useCallback(() => store.getState(), [store]),
-          (prev, next) => Counter.compare(prev, next),
-        )
-      },
-    ],
-    [
-      "memoized watcher with memoized comparator",
-      ({ store }: WithStore) => {
-        return useInnerWatch(
-          useCallback(() => store.getState(), [store]),
-          Counter.compare,
-        )
-      },
-    ],
-  ])("direct %s", (_, hook) => {
-    it.concurrent("watches the store's changes", () => {
-      const store = InnerStore.of({ count: 1 })
-
-      const { result } = renderHook(hook, {
-        initialProps: { store },
-      })
-
-      expect(result.current).toStrictEqual({ count: 1 })
-
-      act(() => {
-        store.setState(Counter.inc)
-      })
-      expect(result.current).toStrictEqual({ count: 2 })
-
-      act(() => {
-        store.setState(({ count }) => ({ count: count * 2 }))
-      })
-      expect(result.current).toStrictEqual({ count: 4 })
-    })
-
-    describe("watches the replaced store changes", () => {
-      const setup = () => {
-        const store_1 = InnerStore.of({ count: 1 })
-        const store_2 = InnerStore.of({ count: 10 })
-
-        const { result, rerender } = renderHook(hook, {
-          initialProps: { store: store_1 },
-        })
-
-        return { store_1, store_2, result, rerender }
-      }
-
-      it.concurrent("initiates with correct result", () => {
-        const { result } = setup()
-
-        expect(result.current).toStrictEqual({ count: 1 })
-      })
-
-      it.concurrent("replaces initial store_1 with store_2", () => {
-        const { store_2, result, rerender } = setup()
-
-        rerender({ store: store_2 })
-        expect(result.current).toStrictEqual({ count: 10 })
-      })
-
-      it.concurrent(
-        "stops watching store_1 changes after replacement with store_2",
-        () => {
-          const { store_1, store_2, result, rerender } = setup()
-
-          rerender({ store: store_2 })
-
-          act(() => {
-            store_1.setState(Counter.inc)
-          })
-
-          expect(store_1.getState()).toStrictEqual({ count: 2 })
-          expect(result.current).toStrictEqual({ count: 10 })
-        },
-      )
-
-      it.concurrent(
-        "starts watching store_2 changes after replacement of store_1",
-        () => {
-          const { store_1, store_2, result, rerender } = setup()
-
-          rerender({ store: store_2 })
-
-          act(() => {
-            store_2.setState(Counter.inc)
-          })
-
-          expect(store_1.getState()).toStrictEqual({ count: 1 })
-          expect(result.current).toStrictEqual({ count: 11 })
-        },
-      )
-
-      it.concurrent("replaces store_1 back", () => {
-        const { store_1, store_2, result, rerender } = setup()
-
-        rerender({ store: store_2 })
-        rerender({ store: store_1 })
-
-        expect(result.current).toStrictEqual({ count: 1 })
-      })
-
-      it.concurrent(
-        "stops watching store_2 after replacement back store_1",
-        () => {
-          const { store_1, store_2, result, rerender } = setup()
-
-          rerender({ store: store_2 })
-          rerender({ store: store_1 })
-
-          act(() => {
-            store_2.setState(Counter.inc)
-          })
-
-          expect(result.current).toStrictEqual({ count: 1 })
-        },
-      )
-    })
-  })
-
-  describe.skip.each([
-    [
-      "inline",
-      ({ store }: WithStore) => {
-        return useInnerWatch(() => store.getState())
-      },
-    ],
-    [
-      "memoized",
-      ({ store }: WithStore) => {
-        return useInnerWatch(useCallback(() => store.getState(), [store]))
-      },
-    ],
-  ])("conditional %s", () => {})
-
-  describe("transform state's value inside watcher", () => {
-    const toTuple = ({ count }: Counter): [boolean, boolean] => {
-      return [count > 2, count < 5]
-    }
-
-    const compareTuple = (
-      [prevLeft, prevRight]: [boolean, boolean],
-      [nextLeft, nextRight]: [boolean, boolean],
-    ): boolean => {
-      return prevLeft === nextLeft && prevRight === nextRight
-    }
-
-    it.concurrent.each([
-      [
-        "inline",
-        ({ store }: WithStore) => {
-          return useInnerWatch(() => store.getState(toTuple))
-        },
-      ],
-      [
-        "memoized",
-        ({ store }: WithStore) => {
-          return useInnerWatch(
-            useCallback(() => store.getState(toTuple), [store]),
-          )
-        },
-      ],
-    ])(
-      "produces new value on each store's update with %s watcher",
-      (_, hook) => {
-        const store = InnerStore.of({ count: 1 })
-
-        const { result, rerender } = renderHook(hook, {
-          initialProps: { store },
-        })
-
-        let prev = result.current
-
-        // produces initial result
-        expect(result.current).toStrictEqual([false, true])
-
-        // increments 1 -> 2
-        prev = result.current
-        act(() => {
-          store.setState(Counter.inc)
-        })
-        expect(result.current).not.toBe(prev)
-        expect(result.current).toStrictEqual([false, true])
-
-        // increments 2 -> 3
-        prev = result.current
-        act(() => {
-          store.setState({ count: 3 })
-        })
-        expect(result.current).not.toBe(prev)
-        expect(result.current).toStrictEqual([true, true])
-
-        // rerender
-        rerender()
-        expect(result.current).toStrictEqual([true, true])
-
-        // increments 3 -> 4
-        prev = result.current
-        act(() => {
-          store.setState({ count: 4 })
-        })
-        expect(result.current).not.toBe(prev)
-        expect(result.current).toStrictEqual([true, true])
-
-        // increments 4 -> 5
-        prev = result.current
-        act(() => {
-          store.setState(Counter.inc)
-        })
-        expect(result.current).not.toBe(prev)
-        expect(result.current).toStrictEqual([true, false])
-      },
-    )
-
-    it.concurrent.each([
-      [
-        "inline watcher with inline comparator",
-        ({ store }: WithStore) => {
-          return useInnerWatch(
-            () => store.getState(toTuple),
-            (prev, next) => compareTuple(prev, next),
-          )
-        },
-      ],
-      [
-        "inline watcher with memoized comparator",
-        ({ store }: WithStore) => {
-          return useInnerWatch(() => store.getState(toTuple), compareTuple)
-        },
-      ],
-      [
-        "memoized watcher with inline comparator",
-        ({ store }: WithStore) => {
-          return useInnerWatch(
-            useCallback(() => store.getState(toTuple), [store]),
-            (prev, next) => compareTuple(prev, next),
-          )
-        },
-      ],
-      [
-        "memoized watcher with memoized comparator",
-        ({ store }: WithStore) => {
-          return useInnerWatch(
-            useCallback(() => store.getState(toTuple), [store]),
-            compareTuple,
-          )
-        },
-      ],
-    ])("keeps the old value when it is comparably equal when %s", (_, hook) => {
-      const store = InnerStore.of({ count: 1 })
-
-      const { result, rerender } = renderHook(hook, {
-        initialProps: { store },
-      })
-
-      let prev = result.current
-
-      // produces initial result
-      expect(result.current).toStrictEqual([false, true])
-
-      // increments 1 -> 2
-      prev = result.current
-      act(() => {
-        store.setState(Counter.inc)
-      })
-      expect(result.current).toBe(prev)
-      expect(result.current).toStrictEqual([false, true])
-
-      // increments 2 -> 3
-      prev = result.current
-      act(() => {
-        store.setState({ count: 3 })
-      })
-      expect(result.current).not.toBe(prev)
-      expect(result.current).toStrictEqual([true, true])
-
-      // rerender
-      rerender()
-      expect(result.current).toStrictEqual([true, true])
-
-      // increments 3 -> 4
-      prev = result.current
-      act(() => {
-        store.setState({ count: 4 })
-      })
-      expect(result.current).toBe(prev)
-      expect(result.current).toStrictEqual([true, true])
-
-      // increments 4 -> 5
-      prev = result.current
-      act(() => {
-        store.setState(Counter.inc)
-      })
-      expect(result.current).not.toBe(prev)
-      expect(result.current).toStrictEqual([true, false])
-    })
-
-    it.concurrent.each([
-      [
-        "inline watcher",
-        ({ spy, store }: WithStore & WithSpy) => {
-          return useInnerWatch(() => {
-            spy()
-
-            return store.getState()
-          })
-        },
-      ],
-      [
-        "inline watcher with inline comparator",
-        ({ spy, store }: WithStore & WithSpy) => {
-          return useInnerWatch(
-            () => {
-              spy()
-
-              return store.getState()
-            },
-            (prev, next) => Counter.compare(prev, next),
-          )
-        },
-      ],
-      [
-        "inline watcher with memoized comparator",
-        ({ spy, store }: WithStore & WithSpy) => {
-          return useInnerWatch(() => {
-            spy()
-
-            return store.getState()
-          }, Counter.compare)
-        },
-      ],
-      [
-        "memoized watcher",
-        ({ spy, store }: WithStore & WithSpy) => {
-          return useInnerWatch(
-            useCallback(() => {
-              spy()
-
-              return store.getState()
-            }, [spy, store]),
-          )
-        },
-      ],
-      [
-        "memoized watcher with inline comparator",
-        ({ spy, store }: WithStore & WithSpy) => {
-          return useInnerWatch(
-            useCallback(() => {
-              spy()
-
-              return store.getState()
-            }, [spy, store]),
-            (prev, next) => Counter.compare(prev, next),
-          )
-        },
-      ],
-      [
-        "memoized watcher with memoized comparator",
-        ({ spy, store }: WithStore & WithSpy) => {
-          return useInnerWatch(
-            useCallback(() => {
-              spy()
-
-              return store.getState()
-            }, [spy, store]),
-            Counter.compare,
-          )
-        },
-      ],
-    ])(
-      "should not trigger the %s when store sets comparably equal state",
-      (_, hook) => {
-        const store = InnerStore.of({ count: 1 })
-        const spy = jest.fn()
-
-        renderHook(hook, {
-          initialProps: { spy, store },
-        })
-
-        // 1st extracts the watcher result
-        // 2nd subscribes to the included stores' changes
-        expect(spy).toHaveBeenCalledTimes(2)
-
-        act(() => {
-          store.setState(Counter.clone, Counter.compare)
-        })
-
-        expect(spy).toHaveBeenCalledTimes(2)
-      },
-    )
-  })
-
-  it.concurrent.each([
-    [
-      "inline watcher",
-      ({ initial, spy, store }: WithStore & WithSpy & WithInitial) => {
-        return useInnerWatch(() => {
-          spy()
-
-          return initial ?? store.getState()
-        })
-      },
-    ],
-    [
-      "inline watcher with inline comparator",
-      ({ initial, spy, store }: WithStore & WithSpy & WithInitial) => {
-        return useInnerWatch(
-          () => {
-            spy()
-
-            return initial ?? store.getState()
-          },
-          (prev, next) => Counter.compare(prev, next),
-        )
-      },
-    ],
-    [
-      "inline watcher with memoized comparator",
-      ({ initial, spy, store }: WithStore & WithSpy & WithInitial) => {
-        return useInnerWatch(() => {
-          spy()
-
-          return initial ?? store.getState()
-        }, Counter.compare)
-      },
-    ],
-    [
-      "memoized watcher",
-      ({ initial, spy, store }: WithStore & WithSpy & WithInitial) => {
-        return useInnerWatch(
-          useCallback(() => {
-            spy()
-
-            return initial ?? store.getState()
-          }, [initial, spy, store]),
-        )
-      },
-    ],
-    [
-      "memoized watcher with inline comparator",
-      ({ initial, spy, store }: WithStore & WithSpy & WithInitial) => {
-        return useInnerWatch(
-          useCallback(() => {
-            spy()
-
-            return initial ?? store.getState()
-          }, [initial, spy, store]),
-          (prev, next) => Counter.compare(prev, next),
-        )
-      },
-    ],
-    [
-      "memoized watcher with memoized comparator",
-      ({ initial, spy, store }: WithStore & WithSpy & WithInitial) => {
-        return useInnerWatch(
-          useCallback(() => {
-            spy()
-
-            return initial ?? store.getState()
-          }, [initial, spy, store]),
-          Counter.compare,
-        )
-      },
-    ],
-  ])(
-    "should not trigger %s when store#getState() is not part of the watcher execution",
-    (_, hook) => {
-      const initial = { count: 0 }
-      const store = InnerStore.of({ count: 1 })
-      const spy = jest.fn()
-
-      const { result } = renderHook(hook, {
-        initialProps: { initial, spy, store },
-      })
-
-      // 1st extracts the watcher result
-      // 2nd subscribes to the included stores' changes
-      expect(spy).toHaveBeenCalledTimes(2)
-      expect(result.current).toBe(initial)
-
-      act(() => {
-        store.setState(Counter.inc)
-      })
-
-      expect(spy).toHaveBeenCalledTimes(2)
-      expect(result.current).toBe(initial)
-    },
-  )
-
-  describe("multiple store#getState() called for the same store", () => {
-    describe.each([
-      [
-        "inline watcher",
-        ({ spy, store }: WithStore & WithSpy, compare?: Compare<Counter>) => {
-          return useInnerWatch(() => {
-            spy()
-
-            return store.getState()
-          }, compare)
-        },
-        ({ spy, store }: WithStore & WithSpy, compare?: Compare<Counter>) => {
-          return useInnerWatch(() => {
-            spy()
-
-            return Counter.merge(store.getState(), store.getState())
-          }, compare)
-        },
-      ],
-      [
-        "memoized watcher",
-        ({ spy, store }: WithStore & WithSpy, compare?: Compare<Counter>) => {
-          return useInnerWatch(
-            useCallback(() => {
-              spy()
-
-              return store.getState()
-            }, [spy, store]),
-            compare,
-          )
-        },
-        ({ spy, store }: WithStore & WithSpy, compare?: Compare<Counter>) => {
-          return useInnerWatch(
-            useCallback(() => {
-              spy()
-
-              return Counter.merge(store.getState(), store.getState())
-            }, [spy, store]),
-            compare,
-          )
-        },
-      ],
-    ])(
-      "triggering %s for multiple store#getState() calls the same as for a single",
-      (_, useSingleHook, useDoubleHook) => {
-        describe.each([
-          ["without comparator", useSingleHook, useDoubleHook],
-          [
-            "with inline comparator",
-            (props: WithStore & WithSpy) => {
-              return useSingleHook(props, (prev, next) =>
-                Counter.compare(prev, next),
-              )
-            },
-            (props: WithStore & WithSpy) => {
-              return useDoubleHook(props, (prev, next) =>
-                Counter.compare(prev, next),
-              )
-            },
-          ],
-          [
-            "with memoized comparator",
-            (props: WithStore & WithSpy) => {
-              return useSingleHook(props, Counter.compare)
-            },
-            (props: WithStore & WithSpy) => {
-              return useDoubleHook(props, Counter.compare)
-            },
-          ],
-        ])("%s", (__, singleHook, doubleHook) => {
-          const setup = () => {
-            const spySingle = jest.fn()
-            const spyDouble = jest.fn()
-            const store = InnerStore.of({ count: 1 })
-
-            const { result: resultSingle } = renderHook(singleHook, {
-              initialProps: { spy: spySingle, store },
-            })
-            const { result: resultDouble } = renderHook(doubleHook, {
-              initialProps: { spy: spyDouble, store },
-            })
-
-            return { store, spySingle, spyDouble, resultSingle, resultDouble }
-          }
-
-          it.concurrent("initiates with expected results", () => {
-            const { spySingle, spyDouble, resultSingle, resultDouble } = setup()
-
-            expect(resultSingle.current).toStrictEqual({ count: 1 })
-            expect(resultDouble.current).toStrictEqual({ count: 2 })
-            expect(spySingle).toHaveBeenCalledTimes(spyDouble.mock.calls.length)
-          })
-
-          it.concurrent.each([
-            // eslint-disable-next-line no-undefined
-            ["without store#setState comparator", undefined],
-            ["with store#setState comparator", Counter.compare],
-          ])("increments %s", (___, compare) => {
-            const { store, spySingle, spyDouble, resultSingle, resultDouble } =
-              setup()
-
-            act(() => {
-              store.setState(Counter.inc, compare)
-            })
-
-            expect(resultSingle.current).toStrictEqual({ count: 2 })
-            expect(resultDouble.current).toStrictEqual({ count: 4 })
-            expect(spySingle).toHaveBeenCalledTimes(spyDouble.mock.calls.length)
-          })
-
-          it.concurrent.each([
-            // eslint-disable-next-line no-undefined
-            ["without store#setState comparator", undefined],
-            ["with store#setState comparator", Counter.compare],
-          ])("clones %s", (___, compare) => {
-            const { store, spySingle, spyDouble, resultSingle, resultDouble } =
-              setup()
-
-            act(() => {
-              store.setState(Counter.clone, compare)
-            })
-
-            expect(resultSingle.current).toStrictEqual({ count: 1 })
-            expect(resultDouble.current).toStrictEqual({ count: 2 })
-            expect(spySingle).toHaveBeenCalledTimes(spyDouble.mock.calls.length)
-          })
-        })
-      },
-    )
-  })
-})
-
 describe("multiple stores are watching", () => {
   describe.each([
     [
